+++ examples/relay-treasurehunt/js/routes/AppHomeRoute.js
@@ -1,12 +1,6 @@
 export default class extends Relay.Route {
   static queries = {
-    game: (Component) => Relay.QL`
-      query {
-        game {
-          ${Component.getFragment('game')},
-        },
-      }
-    `,
+    game: () => Relay.QL`query { game }`,
   };
   static routeName = 'AppHomeRoute';
 }