+++ examples/todo/js/routes/TodoAppHomeRoute.js
@@ -1,12 +1,6 @@
 export default class extends Relay.Route {
   static queries = {
-    viewer: (Component) => Relay.QL`
-      query RootQuery {
-        viewer {
-          ${Component.getFragment('viewer')},
-        },
-      }
-    `,
+    viewer: () => Relay.QL`query RootQuery { viewer }`,
   };
   static routeName = 'TodosHomeRoute';
 }